+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Increment `@folio/stripes` to `v5`, `react-router` to `v5.2`.
 * New developer-setting page: "Okapi query" lets you, you know, query Okapi.
 * Update to react-intl v5. UID-25.
+* Import supported locales from `@folio/stripes/core`.
 
 ## [3.1.0](https://github.com/folio-org/ui-developer/tree/v3.1.0) (IN PROGRESS)
 [Full Changelog](https://github.com/folio-org/ui-developer/compare/v3.0.0...v3.1.0)